requirements.txt
mysql-connector-python==8.1.0
python-dotenv==1.0.0